+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Change Log
 ==========
 
+[dev-master] - 2016-06-17
+-------------------------
+
+### Changed
+- Default handler is now optional
+- `pushHandler` method renamed to `push`
+- `unshiftHandler` method renamed to `unshift`
+- `popHandler` method renamed to `pop`
+- `shiftHandler` method renamed to `shift`
+
 [dev-master] - 2016-05-21
 ---------------------